+++ src/services/ui/src/components/Opensearch/Filtering/consts.ts
@@ -1,6 +1,14 @@
-import { OsField, OsFilterable } from "shared-types";
-import { OsFilterComponentType } from "./types";
+import {
+  OsExportHeaderOptions,
+  OsField,
+  OsFilterable,
+  OsMainSourceItem,
+} from "shared-types";
+import { OsFilterComponentType, OsTab } from "../types";
 import { UserRoles } from "shared-types";
+import { BLANK_VALUE } from "@/consts";
+import { LABELS } from "@/lib/labels";
+import { format } from "date-fns";
 
 type DrawerFilterableGroup = {
   label: string;
@@ -44,9 +52,16 @@ export const FILTER_GROUPS = (
       type: "terms",
       value: [],
     },
-    [(user?.isCms && !user.user?.["custom:cms-roles"].includes(UserRoles.HELPDESK)) ? "cmsStatus.keyword" : "stateStatus.keyword"]: {
+    [user?.isCms &&
+    !user.user?.["custom:cms-roles"].includes(UserRoles.HELPDESK)
+      ? "cmsStatus.keyword"
+      : "stateStatus.keyword"]: {
       label: "Status",
-      field: (user?.isCms && !user.user?.["custom:cms-roles"].includes(UserRoles.HELPDESK)) ? "cmsStatus.keyword" : "stateStatus.keyword",
+      field:
+        user?.isCms &&
+        !user.user?.["custom:cms-roles"].includes(UserRoles.HELPDESK)
+          ? "cmsStatus.keyword"
+          : "stateStatus.keyword",
       component: "multiCheck",
       prefix: "must",
       type: "terms",
@@ -77,4 +92,82 @@ export const FILTER_GROUPS = (
       value: [],
     },
   };
+};
+
+export const EXPORT_GROUPS = (
+  tab: OsTab,
+  user?: any
+): OsExportHeaderOptions<OsMainSourceItem>[] => {
+  return [
+    {
+      name: (() => {
+        if (tab === "spas") {
+          return "SPA ID";
+        } else if (tab === "waivers") {
+          return "Waiver Number";
+        }
+        return "";
+      })(),
+      transform: (data) => data.id,
+    },
+    {
+      name: "State",
+      transform: (data) => data.state ?? BLANK_VALUE,
+    },
+    {
+      name: "Type",
+      transform: (data) => data.planType ?? BLANK_VALUE,
+    },
+    {
+      name: "Action Type",
+      transform: (data) => {
+        if (data.actionType === undefined) {
+          return BLANK_VALUE;
+        }
+
+        return (
+          LABELS[data.actionType as keyof typeof LABELS] || data.actionType
+        );
+      },
+    },
+    {
+      name: "Status",
+      transform(data) {
+        if (user?.data?.isCms && !user?.data?.user) {
+          if (data.cmsStatus) {
+            return data.cmsStatus;
+          }
+          return BLANK_VALUE;
+        } else {
+          if (data.stateStatus) {
+            return data.stateStatus;
+          }
+          return BLANK_VALUE;
+        }
+      },
+    },
+    {
+      name: "Initial Submission",
+      transform: (data) =>
+        data?.submissionDate
+          ? format(new Date(data.submissionDate), "MM/dd/yyyy")
+          : BLANK_VALUE,
+    },
+    {
+      name: "Formal RAI Response",
+      transform: (data) => {
+        return data.raiReceivedDate
+          ? format(new Date(data.raiReceivedDate), "MM/dd/yyyy")
+          : BLANK_VALUE;
+      },
+    },
+    {
+      name: "CPOC Name",
+      transform: (data) => data.leadAnalystName ?? BLANK_VALUE,
+    },
+    {
+      name: "Submitted By",
+      transform: (data) => data.submitterName ?? BLANK_VALUE,
+    },
+  ];
 };
