+++ src/services/ui/src/components/Opensearch/Filtering/index.tsx
@@ -6,9 +6,7 @@ import { useOsContext } from "../Provider";
 import { OsFilterDrawer } from "./FilterDrawer";
 import { getAllSearchData } from "@/api";
 import { useGetUser } from "@/api/useGetUser";
-import { BLANK_VALUE } from "@/consts";
-import { format } from "date-fns";
-import { LABELS } from "@/lib/labels";
+import { EXPORT_GROUPS } from "./consts";
 
 export const OsFiltering: FC<{
   disabled?: boolean;
@@ -38,79 +36,7 @@ export const OsFiltering: FC<{
         <div className="flex flex-row gap-2">
           <ExportButton
             data={() => getAllSearchData([...params.state.filters, ...filters])}
-            headers={[
-              {
-                name: (() => {
-                  if (params.state.tab === "spas") {
-                    return "SPA ID";
-                  } else if (params.state.tab === "waivers") {
-                    return "Waiver Number";
-                  }
-                  return "";
-                })(),
-                transform: (data) => data.id,
-              },
-              {
-                name: "State",
-                transform: (data) => data.state ?? BLANK_VALUE,
-              },
-              {
-                name: "Type",
-                transform: (data) => data.planType ?? BLANK_VALUE,
-              },
-              {
-                name: "Action Type",
-                transform: (data) => {
-                  if (data.actionType === undefined) {
-                    return BLANK_VALUE;
-                  }
-
-                  return (
-                    LABELS[data.actionType as keyof typeof LABELS] ||
-                    data.actionType
-                  );
-                },
-              },
-              {
-                name: "Status",
-                transform(data) {
-                  if (user?.data?.isCms && !user?.data?.user) {
-                    if (data.cmsStatus) {
-                      return data.cmsStatus;
-                    }
-                    return BLANK_VALUE;
-                  } else {
-                    if (data.stateStatus) {
-                      return data.stateStatus;
-                    }
-                    return BLANK_VALUE;
-                  }
-                },
-              },
-              {
-                name: "Initial Submission",
-                transform: (data) =>
-                  data?.submissionDate
-                    ? format(new Date(data.submissionDate), "MM/dd/yyyy")
-                    : BLANK_VALUE,
-              },
-              {
-                name: "Formal RAI Response",
-                transform: (data) => {
-                  return data.raiReceivedDate
-                    ? format(new Date(data.raiReceivedDate), "MM/dd/yyyy")
-                    : BLANK_VALUE;
-                },
-              },
-              {
-                name: "CPOC Name",
-                transform: (data) => data.leadAnalystName ?? BLANK_VALUE,
-              },
-              {
-                name: "Submitted By",
-                transform: (data) => data.submitterName ?? BLANK_VALUE,
-              },
-            ]}
+            headers={EXPORT_GROUPS(params.state.tab, user)}
           />
           <OsFilterDrawer />
         </div>
