+++ src/services/ui/src/components/Opensearch/useOpensearch.ts
@@ -5,7 +5,8 @@ import { OsQueryState, SearchData, UserRoles } from "shared-types";
 import { createSearchFilterable } from "./utils";
 import { useQuery } from "@tanstack/react-query";
 import { useGetUser } from "@/api/useGetUser";
-export type OsTab = "waivers" | "spas";
+import { OsTab } from "./types";
+
 export const DEFAULT_FILTERS: Record<OsTab, Partial<OsParamsState>> = {
   spas: {
     filters: [
@@ -91,8 +92,16 @@ export const useOsAggregate = () => {
             size: 10,
           },
           {
-            field: (user?.isCms && !user.user?.["custom:cms-roles"].includes(UserRoles.HELPDESK)) ? "cmsStatus.keyword" : "stateStatus.keyword",
-            name: (user?.isCms && !user.user?.["custom:cms-roles"].includes(UserRoles.HELPDESK)) ? "cmsStatus.keyword" : "stateStatus.keyword",
+            field:
+              user?.isCms &&
+              !user.user?.["custom:cms-roles"].includes(UserRoles.HELPDESK)
+                ? "cmsStatus.keyword"
+                : "stateStatus.keyword",
+            name:
+              user?.isCms &&
+              !user.user?.["custom:cms-roles"].includes(UserRoles.HELPDESK)
+                ? "cmsStatus.keyword"
+                : "stateStatus.keyword",
             type: "terms",
             size: 10,
           },
